requirements.txt
sleipnirgroup-jormungandr
casadi
matplotlib
scicy